Move logger shim out of alignment script
Replace the generated @elizaos/logger string payload with a tracked shim package under scripts/shims. The alignment script now links that package directly, avoiding the oversized helper, generated template strings, and dynamic file writes Codacy reported.
