+++ scripts/shims/elizaos-logger/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "@elizaos/logger",
+  "version": "0.0.0-milady-local",
+  "type": "module",
+  "main": "./index.js",
+  "module": "./index.js",
+  "types": "./index.d.ts",
+  "exports": {
+    ".": {
+      "types": "./index.d.ts",
+      "import": "./index.js",
+      "default": "./index.js"
+    }
+  }
+}
